data/yt-sb-1b/video2dataset-1.1.0/requirements.txt
numpy>=1.19.5,<2
fire>=0.4.0,<0.5.0
requests>=2.27.1,<3
ffmpeg-python
yt_dlp
pyarrow
fsspec
webdataset
webvtt-py
wandb
pandas
tqdm
timeout-decorator
scenedetect[opencv]
decord
torch>=1.7.1,<2
langdetect